requirements.txt
absl-py==0.15.0
Cython==0.29.26
matplotlib>=3.0.3
numpy==1.22.0
opencv-python>=4.6.0
Pillow==9.3.0
protobuf==3.20.2
PyYAML==6.0
six>=1.15.0
tensorflow==2.9.3
tensorflow-addons>=0.12
tensorflow-hub>=0.11
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-model-optimization==0.7.2

